# platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDeskViewer RustDesk Pro API
  version: 1.0.0
  description: |
    RustDesk Pro-compatible API for OpenDeskViewer management platform.
    This API is consumed by the unmodified RustDesk client.
    
    All endpoints use snake_case and return `{"error": "..."}` on failure.
    Pagination uses `?current=&pageSize=` with response `{total, data}`.
    Authentication: `Authorization: Bearer <token>` for authenticated endpoints.
servers:
  - url: http://localhost:8000
paths:
...
  /api/oidc/callback:
    get:
      summary: Where Keycloak sends the browser
      description: >-
        Not called by the client. Verifies the state against a sign-in this
        server started, exchanges the authorization code with PKCE, validates
        the returned token against the same issuer and audience every other
        route requires, and records which user authenticated. Renders a page
        telling the user to return to the application; it carries no token.
      operationId: getOidcCallback
      tags:
        - OIDC
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        '200':
          description: An HTML page, sign-in complete or cancelled
          content:
            text/html:
              schema:
                type: string
        '400':
          description: A missing, unknown or expired state
        '401':
          description: The token returned by the identity provider failed validation
        '502':
          description: The identity provider refused the exchange

  # ========== Address Book (ab) ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...